requirements.txt
# Core dependencies
PyYAML>=6.0

# Optional dependencies for data loading
pandas>=1.5.0  # For Parquet file support
pyarrow>=10.0.0  # For Parquet file support
datasets>=2.14.0  # For Hugging Face dataset support

# Optional: Symbolic math (for advanced calculator features)
# sympy>=1.12  # Uncomment if needed for symbolic computation

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
